+++ WpfKeyboardSimulator/WpfKeyboardSimulatorApp/logic/IUserInput.cs
@@ -1,0 +1,7 @@
+﻿namespace WpfKeyboardSimulatorApp.model;
+
+public interface IUserInput
+{
+    void SetUpperCase();
+    void SetLowerCase();
+}
