+++ WpfKeyboardSimulator/WpfKeyboardSimulatorApp/model/LetterCase.cs
@@ -1,0 +1,6 @@
+﻿namespace WpfKeyboardSimulatorApp.model;
+
+public enum LetterCase
+{
+    Upper, Lower
+}
